Upto step5 -pn ok
I had to determine the right clustering file _while_ step05 is running
and not with a params otherwise it won't be able to find step04 results.
The R code looks in the clustering file and determines the n. of
clusters. Do we need to check that this corresponds to the path
'declared' n. of clusters? Hope not.